+++ druid/src/widget/composite.rs
@@ -6,13 +6,15 @@ use crate::{
     UpdateCtx, Widget, WidgetPod,
 };
 
+/// Meta information for Widget derive
 pub struct CompositeMeta<T> {
-    pub child: Option<WidgetPod<T, Box<dyn Widget<T>>>>,
+    /// Built widget
+    pub widget: Option<WidgetPod<T, Box<dyn Widget<T>>>>,
 }
 
 impl<T> Default for CompositeMeta<T> {
     fn default() -> Self {
-        CompositeMeta { child: None }
+        CompositeMeta { widget: None }
     }
 }
